src/connector/requirements.txt
nats-py
python-dotenv
protobuf
